Add strict param to OsvSources
This is useful for schema evolution. For example if we add a new field in a ScaldingType model we will be able to use the new code to read older files. In this case the value of the new field will be the default set in the ScaldingType or null in case there is no default.